Cleaned up search results
Fixed and cleaned up an issue where if the MusicBrainz API didn't have any values for certain artists or releases, the values would return as null. This has been fixed now where it will check if the values are null and if they are, update the text to say "Unknown".