Fixed bugs related to search
Fixed two major bugs related to searching for an artist. the search would break if you searched for a band compared to a single artist because they don't always have a "date of birth" but a year they were formed, so I added some if statements to check if the data returned is a year only, and if so, adjust the output accordingly. The other bug was a similar issue but with the gender of an artist. If they had no gender listed in the data, the attempt to use substring would break the search result. That was also fixed by adding an if statement to only use the substring if it detects the gender field isn't null.